renamed query model file, added email to createUser
