--- server/routes/loginRouter.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const loginController = require('../controllers/loginController');
-
-router.get('/:username/:password', loginController.checkUser);
-
-router.get('/logout', loginController.logout);
-
-module.exports = router;
